% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot_compare_all}
\alias{boot_compare_all}
\title{Apply a two group statistical test to each group in a nested tibble}
\usage{
boot_compare_all(dat, statistic, within_group = NULL, reps = 10000,
  ...)
}
\description{
Apply a two group statistical test to each group in a nested tibble
}
