% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootr.r
\name{bootr.get_parameters}
\alias{bootr.get_parameters}
\title{Get estimates from a bootstrap run.}
\usage{
bootr.get_parameters(i, bout, type = "perc")
}
\arguments{
\item{i}{The index from the bootstrap run.}

\item{bout}{The return from the boostrap run.}

\item{type}{The type of confidence interval to generate (defaults to "perc")}
}
\description{
This function allows you to get the estimates from a bootstrap run.
}
\examples{
bootr.get_parameters()
}
\keyword{bootstrap}
