% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootr.r
\name{bootr.predictions}
\alias{bootr.predictions}
\title{Get predictions from a model}
\usage{
bootr.predictions(new_data, model, full_data, func = predict,
  weighted_mean = NULL, .labels = T, ...)
}
\arguments{
\item{new_data}{A dataframe that list the values to produce predictions at.}

\item{model}{The model from which predictions are produced.}

\item{full_data}{The full dataset used to run the model.}

\item{func}{The function to produce predictions (defaults to "predict")}

\item{weighted_mean}{A vector of observation weights (defaults to "NULL")}
}
\description{
This function allows you to generate predictions from a model using the observed values approach.
}
\examples{
bootr.predictions()
}
\keyword{bootstrap}
\keyword{predict,}
