% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootr.R
\name{bootr}
\alias{bootr}
\title{Bootstrap calculations on a data frame}
\usage{
bootr(df, m = 100, group_vars, source_var, boot_fun = mean, ...)
}
\arguments{
\item{df}{A data frame}

\item{m}{Number of replications (Defaults to 100)}

\item{group_vars}{A character vector of variable names by which to group}

\item{source_var}{A single column name on which to calculate the summary
statistic.}

\item{boot_fun}{A summary function (defaults to \link{\code{mean}}) that will
be used to calculate the boostrapped statistic}

\item{...}{Additional arugments passed to \code{boot_fun}}
}
\value{
A summary data frame with the columns \code{group_vars} and
  \code{boot_low} (2.5th percentile), \code{boot_med} (50th percentile), and
  \code{boot_high} (97.5th percentile).
}
\description{
Bootstrap calculations on a data frame
}

