% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootr_output.r
\name{bootr.summary_stats}
\alias{bootr.summary_stats}
\title{Create summary stats for a dataframe.}
\usage{
bootr.summary_stats(df, variable_names)
}
\arguments{
\item{df}{The dataframe to create summary stats for.}
}
\description{
This function allows you to create summary stats for a dataframe.
}
\examples{
bootr.summary_stats()
}
\keyword{stats}
\keyword{summary}

