% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootr_output.r
\name{bootr.write}
\alias{bootr.write}
\title{Write the output of a bootstrap run.}
\usage{
bootr.write(models, decimals = 3, variable_labels = NULL,
  additional = NULL, intercept = "(Intercept)", filename = NULL)
}
\arguments{
\item{models}{The list of models to write out.}
}
\description{
This function allows you to write the output of a bootstrap run to an HTML file.
}
\examples{
bootr.write()
}
\keyword{bootstrap}
